Replaced double with BigDecimal

+++ backend/src/main/java/com/github/redstylzz/backend/model/dto/PaymentDTO.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -14,6 +15,6 @@ public class PaymentDTO {
     String paymentID;
     String categoryID;
     String description;
-    double amount;
+    BigDecimal amount;
     Date payDate;
 }
